prompt_check
