+++ src/errors.h
@@ -13,6 +13,8 @@
 #ifndef ERRORS_H
 # define ERRORS_H
 
-# define ERR_QUOTES "syntax error, quotes not completed"
+// need to check the right error_msg
+# define ERR_QUOTES "syntax error: quotes not completed"
+# define ERR_SYNTAX "syntax error: parse error near "
 
 #endif
